feat: Implement campaign editing
