+++ lib/types.ts
@@ -9,3 +9,25 @@ export interface CampaignInput {
   image?: string;
   isActive: boolean;
 }
+
+export interface ProductResourceItem {
+  id: string;
+  handle: string;
+  title: string;
+  images: {
+    src: string;
+  }[];
+  variants: {
+    id: string;
+  }[];
+}
+
+export interface CollectionResourceItem {
+  id: string;
+  handle: string;
+  title: string;
+  image?: {
+    src: string;
+  };
+  productsCount: number;
+}
