+++ next.config.js
@@ -13,6 +13,16 @@ const nextConfig = {
     CONFIG_SHOPIFY_API_KEY: process.env.SHOPIFY_API_KEY,
     CONFIG_SHOPIFY_APP_URL: process.env.SHOPIFY_APP_URL,
   },
+  experimental: {
+    swcPlugins: [
+      [
+        "next-superjson-plugin",
+        {
+          excluded: [],
+        },
+      ],
+    ],
+  },
 };
 
 export default nextConfig;
