+++ package.json
@@ -45,11 +45,13 @@
     "next": "13.4.10",
     "next-api-middleware": "^2.0.1",
     "next-connect": "^1.0.0",
+    "next-superjson-plugin": "^0.5.9",
     "postcss": "^8.4.27",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.45.2",
     "react-query": "^3.39.3",
+    "superjson": "^1.13.1",
     "tailwindcss": "^3.3.3",
     "yup": "^1.2.0"
   },
