+++ pages/api/auth/callback.js
@@ -23,7 +23,7 @@ const handler = async (req, res) => {
     await prisma.stores.upsert({
       where: { shop: shop },
       update: { isActive: true },
-      create: { shop: shop, identifier: shop, isActive: true },
+      create: { shop: shop, isActive: true },
     });
 
     // Redirect to app with shop parameter upon auth
